requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
defusedxml==0.8.0rc2
dj-database-url==2.2.0
Django==4.2.16
django-allauth==65.0.1
djangorestframework==3.15.2
idna==3.7
mysqlclient==2.2.4
oauthlib==3.2.2
pillow==10.4.0
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.9.0
python-decouple==3.8
python3-openid==3.2.0
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
sqlparse==0.5.1
stripe==10.11.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
